--- features/select-doc/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/selectdoc/internal/drivinglicence/SelectDrivingLicenseState.kt
@@ -1,5 +1,0 @@
-package uk.gov.onelogin.criorchestrator.features.selectdoc.internal.drivinglicence
-
-data class SelectDrivingLicenseState(
-    val displayReadMoreButton: Boolean,
-)
