+++ features/select-doc/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/selectdoc/internal/drivinglicence/select/SelectDrivingLicenceScreen.kt
@@ -32,7 +32,7 @@ import uk.gov.android.ui.patterns.leftalignedscreen.LeftAlignedScreen
 import uk.gov.android.ui.theme.m3.GdsTheme
 import uk.gov.android.ui.theme.util.UnstableDesignSystemAPI
 import uk.gov.onelogin.criorchestrator.features.selectdoc.internal.R
-import uk.gov.onelogin.criorchestrator.features.selectdoc.internalapi.nav.SelectDocumentDestinations
+import uk.gov.onelogin.criorchestrator.features.selectdoc.internalapi.nav.SelectDocDestinations
 import uk.gov.onelogin.criorchestrator.libraries.composeutils.LightDarkBothLocalesPreview
 
 @Composable
@@ -50,22 +50,22 @@ internal fun SelectDrivingLicenceScreen(
             when (event) {
                 SelectDrivingLicenceAction.NavigateToConfirmNoChippedID ->
                     navController.navigate(
-                        SelectDocumentDestinations.ConfirmNoChippedID,
+                        SelectDocDestinations.ConfirmNoChippedID,
                     )
 
                 SelectDrivingLicenceAction.NavigateToConfirmNoNonChippedID ->
                     navController.navigate(
-                        SelectDocumentDestinations.ConfirmNoNonChippedID,
+                        SelectDocDestinations.ConfirmNoNonChippedID,
                     )
 
                 SelectDrivingLicenceAction.NavigateToConfirmation ->
                     navController.navigate(
-                        SelectDocumentDestinations.ConfirmDrivingLicence,
+                        SelectDocDestinations.ConfirmDrivingLicence,
                     )
 
                 SelectDrivingLicenceAction.NavigateToTypesOfPhotoID ->
                     navController.navigate(
-                        SelectDocumentDestinations.TypesOfPhotoID,
+                        SelectDocDestinations.TypesOfPhotoID,
                     )
             }
         }
