+++ features/select-doc/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/selectdoc/internal/drivinglicence/select/SelectDrivingLicenseState.kt
@@ -1,0 +1,5 @@
+package uk.gov.onelogin.criorchestrator.features.selectdoc.internal.drivinglicence.select
+
+data class SelectDrivingLicenseState(
+    val displayReadMoreButton: Boolean,
+)
